requirements.txt
streamlit
Pillow
easyocr
Levenshtein
Unidecode
